services/api/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI & ML
httpx==0.26.0
openai==1.10.0
edge-tts==6.1.9
Pillow==10.2.0

# Storage
boto3==1.34.0
aiobotocore==2.11.0

# Queue & Cache
redis==5.0.1
celery==5.3.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
python-slugify==8.0.1
maxminddb==2.2.0
geoip2==4.2.0

# Payments
razorpay==1.4.1

# Monitoring
sentry-sdk[fastapi]==1.39.2